- setup terraform and update deployment readme
