+++ deployment/config/dev.tfvars.template
@@ -1,0 +1,3 @@
+aws_secret_key  = "${AWS_SECRET_ACCESS_KEY}"
+aws_access_key= "${AWS_ACCESS_KEY}"
+region = "${AWS_REGION}"
