--- deployment/docker.tf
@@ -1,14 +1,0 @@
-#resource "docker_image" "zoo" {
-#  name = "zoo"
-#  build {
-#    context = "."
-#    tag     = ["zoo:develop"]
-#    build_arg = {
-#      foo : "zoo"
-#    }
-#    label = {
-#      author : "zoo"
-#    }
-#  }
-#}
-
