--- deployment/ecr.tf
@@ -1,3 +1,0 @@
-resource "aws_ecr_repository" "ecr_repo" {
-  name = "ecr_afri_repo"
-}
