+++ deployment/old_files/ecr.tf
@@ -1,0 +1,3 @@
+#resource "aws_ecr_repository" "ecr_repo" {
+#  name = "ecr_afri_repo"
+#}
