+++ deployment/old_files/source/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "signed-url",
+  "version": "1.0.0",
+  "description": "",
+  "type": "module",
+  "main": "signedUrl.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "aws-sdk": "^2.1359.0"
+  }
+}
