--- deployment/prod.tfvars.template
@@ -1,3 +1,0 @@
-aws_secret_key  = "${AWS_SECRET_ACCESS_KEY}"
-aws_access_key= "${AWS_ACCESS_KEY}"
-region = "X"
