+++ deployment/required_providers.tf
@@ -1,0 +1,13 @@
+terraform {
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 4.0"
+    }
+    docker = {
+      source = "kreuzwerker/docker"
+      version = "3.0.1"
+    }
+  }
+  required_version = ">= 0.14"
+}
