--- deployment/variables.tf
@@ -1,33 +1,0 @@
-variable "aws_access_key" {
-  type        = string
-  description = "Service account credentials access key"
-}
-
-variable "aws_secret_key" {
-  type        = string
-  description = "service account credentials secret key"
-}
-
-variable "region" {
-  type = string
-  description = "AWS Region"
-}
-
-variable "account_id" {
-  type = string
-  description = " Account id"
-}
-
-variable "put_signedUrl_endpoint_path" {
-  type = string
-  description = "THe endpoint path for generating signed Url"
-  default = "generate_signedUrl"
-}
-
-
-variable "get_object_signedUrl_endpoint_path" {
-  type = string
-  description = "THe endpoint path for generating signed Url"
-  default = "get_object_signed_url"
-}
-
