--- deployment/version.tf
@@ -1,13 +1,0 @@
-terraform {
-  required_providers {
-    aws = {
-      source  = "hashicorp/aws"
-      version = "~> 4.0"
-    }
-    docker = {
-      source = "kreuzwerker/docker"
-      version = "3.0.1"
-    }
-  }
-  required_version = ">= 0.14"
-}
